added pets gallery

+++ api/upload.php
@@ -8,42 +8,47 @@ require "../vendor/autoload.php"; // loads the libraries
 use Cloudinary\Cloudinary; //using the Cloudinary library for PHP
 
 
-if (!isset($_FILES["myFile"]) || $_FILES["myFile"]["size"] === 0) { //if the global var is not set -> return error message
+$files = $_FILES['file'];
+$catch = array();
+
+$allowedTypes = [
+    'image/png' => 'png',
+    'image/jpeg' => 'jpg',
+];
+if (count($files['name']) > 5){
+    die("5");
+}
+if (!isset($_FILES["file"]) || $_FILES["file"]["size"] === 0) { //if the global var is not set -> return error message
     die("1");
 }
 
-$filepath = $_FILES['myFile']['tmp_name']; 
-$fileSize = filesize($filepath);
+for($i = 0;$i < count($files['name']);$i++){
+    $filepath = $files['tmp_name'][$i];
+    $fileSize = filesize($filepath);
 
-$fileinfo = finfo_open(FILEINFO_MIME_TYPE);
-$filetype = finfo_file($fileinfo, $filepath);
+    $fileinfo = finfo_open(FILEINFO_MIME_TYPE);
+    $filetype = finfo_file($fileinfo, $filepath);
 
+    if($fileSize > 3146728){
+        die("2");
+    }
 
+    if (!in_array($filetype, array_keys($allowedTypes))) {
+        die("3");
+    }
 
-if ($fileSize > 3145728) { // 3 MB (1 byte * 1024 * 1024 * 3 (for 3 MB))
-    die("2");
-}
+    $filename = basename($filepath); // use the original name of the file
+    $extension = $allowedTypes[$filetype];
+    $targetDirectory = __DIR__ . "/"; // __DIR__ is the directory of the current PHP file
 
-$allowedTypes = [
-   'image/png' => 'png',
-   'image/jpeg' => 'jpg',
-];
+    $newFilepath = $targetDirectory . "../assets/img/" . $filename . "." . $extension;
 
-if (!in_array($filetype, array_keys($allowedTypes))) {
-    die("3");
-}
+    if (!copy($filepath, $newFilepath)) { // Copy the file, returns false if failed
+        die("4");
+    }
 
-$filename = basename($filepath); // use the original name of the file
-$extension = $allowedTypes[$filetype];
-$targetDirectory = __DIR__ . "/"; // __DIR__ is the directory of the current PHP file
+    unlink($filepath); // Delete the temp file
 
-$newFilepath = $targetDirectory . "../assets/img/" . $filename . "." . $extension;
-
-if (!copy($filepath, $newFilepath)) { // Copy the file, returns false if failed
-    die("4");
-}
-
-unlink($filepath); // Delete the temp file
 
 $cloud_name = getenv('CLOUD_NAME');
 $api_key = getenv('API_KEY');
@@ -51,27 +56,30 @@ $api_secret = getenv('API_SECRET');
 
 
 
-$cloudinary = new Cloudinary([
+// TODO no need to create new Cloudnary object every time
+
+    $cloudinary = new Cloudinary([
     'cloud' => [
         'cloud_name' => $cloud_name,
         'api_key'    => $api_key,
         'api_secret' => $api_secret,
     ],
-]);
+    ]);
 
 
-$resolved_path = "../assets/img/".$filename.".".$extension;
+    $resolved_path = "../assets/img/".$filename.".".$extension;
 
 
-$catch = $cloudinary->uploadApi()->upload($resolved_path); // perform the upload of the picture
-// returns an json object
+    $temp = $cloudinary->uploadApi()->upload($resolved_path); // perform the upload of the picture
+    $catch[$i] = $temp['secure_url'];
+
+    //json_encode(array("secure_url" => $temp['secure_url']),JSON_UNESCAPED_SLASHES);
+    
+    // delete the temporary file on the server side
+    unlink($newFilepath);
+}
 
 
-// return the secure url to the ajax call
-
-print_r(json_encode(array("secure_url" => $catch['secure_url']),JSON_UNESCAPED_SLASHES));
-// delete the temporary file on the server side
-unlink($newFilepath);
-
+print_r(json_encode($catch, JSON_FORCE_OBJECT | JSON_UNESCAPED_SLASHES)."\n");
 
 ?>
